prompt variables, conda, comments
Added an initial 'proof of concept' for the script to show how prompt variables may be used to generate better outputs. 

As of now, starting your prompt with "#Q" will define your prompt as a question and will start the prompt with with the text: "Trending Question: " and end it with "Top Rated Answer: "

This will allow for easier question based prompting. This specific prompt variable may change and/or new prompt variables will be added if and when better prompting methods are discovered.

Incorporated conda environment into script for ease of use. My neox conda environment list can be found in the neoPrompt repository

more comments have been added to explain certain lines :)